backend/requirements.txt
fastapi
uvicorn[standard]
pydantic
python-multipart
openai
google-generativeai
pypdf
pdfplumber
python-docx
supabase